refactor: 랭킹 조회 시 index값+ 페이지네이션 으로 랭킹 순위 부여하도록 수정
