+++ package.json
@@ -33,6 +33,7 @@
     "styled-reset": "^4.5.2"
   },
   "devDependencies": {
+    "@types/lodash": "^4.17.19",
     "@types/node": "^22.13.1",
     "@types/react": "^18.3.3",
     "@types/react-dom": "^18.3.0",
